Improve Trip Info
Click station markers to toggle between displaying trips starting at
that station, trips ending at that station, or remove trips display.

Displayed trips appear as pink arrows on the map starting at one station
and ending at another. The thickness of the arrow depends on how many
trips there are that match that start and end station.

Click on a trip arrow to display a popup with aggregated trip data.